ProxyShape::onPrimResync: clarify naming
the function should always either always operate on the passed in arg (previousPrims) or the class member (m_variantSwitchedPrims), not a mixture of both. This makes it use only the arg (previousPrims).
Functionally, there is no difference, since onPrimResync is only ever called with previousPrims = m_variantSwitchedPrims, but this makes the function more readable