Revert "fix potential crash due to unique_ptr not being initialized to nullptr"
This reverts commit d7b3f645f466a07dbf973dbf523329345164ebba.

This fix was a red herring - maybe I just got lucky the few times I tried this,
and it seemed to fix things. Real issue is that we are doing double extracts
from UsdImagingValueCache on transform and extent values; this happens because
HdMayaDelegateCtx::FitFrustumToRprims queries transform values to calculate a
good frustrum for it's shadow projection matrix; this means that
UsdImagingDelegate.GetTransform / GetExtent may be called both here, and when
rendering the transform itself.

However, it seems odd that you're not allowed to call GetTransform / GetExtent
more than once - especially since this is a publically exposed override of
HdSceneDelegate::GetTransform. Made a post on usd-interest here about this:

https://groups.google.com/forum/#!topic/usd-interest/tIEp5UyDpxs

Seems like the fix should be in usd itself, as there doesn't seem to be a "safe"
/ non-destructive way to get at this information currently...