refactor to primarily use an HdMayaAdapter, not an HdMayaDelegate
This has gives us nice tracking of various dag-level stuff - for instance, will
be much easier to implement visibility toggling, and movement when parent
xforms are moved.

Also gives us better handling of instances, for relatively "free", since
adapters are per-instance.

However, we're still keeping around the delegate for now, for
PopulateSelectedPaths and PreFrame:

PopulateSelectedPaths:
We keep this to deal with selection highlighting when an AL_usdmaya_Transform
is selected.  In these situations, the HdMayaSceneDelegate + Adapter system
doesn't really work, because it iterates over leaf-level shapes in the selection
(which is the first problem, because it's a selected transform, with no shape
underneath), and then looks up the adapter (which is the second problem -
we have no adapter for AL_usdmaya_Transform, and even if we did, HdMaya makes
assumptions that we never need adapters for transforms, just shapes).
It seemed easier to deal with this by just using the HdMayaALProxyDelegate to
handle PopulateSelectedPaths, somewhat similar to how it did previously.

PreFrame:
Currently, HdMayaSceneDelegate has no facility to for adapters to customize the
PreFrame step.  We could add a virtual method (which by default does nothing),
but it would be pretty wasteful to have to iterate over ALL adpaters, just to
find the few that did do something.  So, to do this "right" within
HdMayaSceneDelegate, we'd probably have to add some way for adapters to register
themselves as having a PreFrame, and then keeping track of these.
For now, seemed easier to just tack this onto our HdMayaALProxyDelegate