use RUNPATH, instead of RPATH
This allows it to be used by different point-version of maya than the one it
was built against.

When it was using RPATH, that takes precedence over LD_LIBRARY_PATH - so even
though maya itself sets LD_LIBRARY_PATH when it starts, this is ignored, and
maya_hydra was still using the libs from the maya it was built against.