MAYA-98188 selection support for USD instancing object (#50)
* MAYA-98188 selection support for USD instancing object

The current selection support for USD instancing object is parity
of USDView as below.

* Precise selection for common instancing.

* Coarse selection for nested instancing:

Only the highest level in the instancing hierarchy will be selected,
which means a group of objects will be selected. If it is important
to support precise selection for nested instancing, we can defer the
work after API freeze because the implementation is on plugin side.