remove Pixar-specific performance test baseline yaml files
The metrics stored in these files are only meaningful in terms of Pixar's
particular hardware/software configuration. Though the performance tests
themselves are available in the repo, the tests are not currently registered
and run through CMake. These tests will continue to be run and tracked against
the baselines internally at Pixar, but otherwise these files are not useful to
anyone else.