sanitize the proxy's UsdImagingDelegate name before using it to construct the delegateId
When a proxy node lives inside a Maya namespace (as is the case when a
pxrUsdReferenceAssembly node is loaded in its "Collapsed" representation), its
name will include colon characters which are invalid in identifier names used
in SdfPaths. This change ensures that the delegate name is a valid identifier
before we try to create an SdfPath from it.