skip fragment registration/de-registration when MRenderer or MFragmentManager are not available
This resolves an issue where headless contexts of Maya would emit API errors
during plugin registration because the renderer cannot be initialized. We now
no longer force the renderer to initialize, and we bypass fragment
registration/de-registration and return success if we cannot get a handle to
the MRenderer or MFragmentManager.