[al] make pushToPrim default to On
...a recent AL change (55e01cf9e52d14069f86763bf447a9c6ee7650a9) added
the ability to control the setting for pushToPrim for newly created
nodes via an optionVar, which is useful, but it also changed the default
from on to off.

In addition to being breaking behavior, this is also highly
non-intuitive for artists, who will be confused when they try to
transform the usd object, and it does nothing.  I would also argue that
the ability to pushToPrim is the primary thing which distinguishes this
plugin from the Pixar plugin, and as such, is probably why most people
are using it in the first place.