move the compiler definition of USD_VERSION_NUM to CXXDefaults.cmake
This lifts the compiler definition of USD_VERSION_NUM that was previously
specified only in AL_USDMaya out of that library and makes it a project-wide
default. Since the value is based on the version of USD being built against, it
will not change between different components in maya-usd. This avoids needing
to specify it as a compiler definition in every component where it is used and
in every component that includes something that uses it.