Handle anonymous layer in ProxyShape
This change allows ProxyShape to correctly load an anonymous root
layer if the file path is set to an anonymous layer id.

Note that the filePath on the ProxyShape is subject to change when
an anonymous layer is used. This is necessary because an
anonymous layer cannot be re-created with the same exact anon
identitifier. This prevents the ProxyShape's file path from deviating
from the layer manager serialisation records, and by doing so
guarantees fetching the right layer when the Maya scene is reloaded.

A couple of Python tests are provided to verify this feature.