add MProfiler and trace performance tracking to core UsdMayaGLBatchRenderer
The MProfilerScopes added are helpful for tracking overall time spent in the
batch renderer when using Maya's Profiler. The trace performance tracking is
good for looking at the time spent in batch renderer setup and overhead versus
the time spent in underlying Hydra and USD libraries, since those are also
instrumented using trace.