identify Qt 5.12 as the minimum required version for Qt-dependent features
The newly added import dialog functionality makes use of QOverload which was
introduced in Qt 5.7, and QSortFilterProxyModel::setRecursiveFilteringEnabled()
which was introduced in Qt 5.10. Maya 2020 itself is built with Qt 5.12.6, so
it is recommended that mayaUsd be built with the same version of Qt.

The Qt features are not supported in Maya 2018 and 2019, since those versions
of Maya use Qt 5.6.1. VFX platform 2018 specifies Qt 5.6.1 as the required
version, so users on that platform may not be able to build the Qt-dependent
features until they switch to VFX platform 2019 which specifies Qt 5.12
instead.

This change adds CMake enforcement of Qt 5.12 when building with Qt enabled,
and adds a table of version numbers for the optional dependencies, of which Qt
is currently the only one.