fix missing geometry in testProxyShapeLiveSurface test
When executing the test script, Maya may reach the point in the test where
we're clicking in the viewport before we've had a chance to evaluate or draw
any of the proxy geometry, since simply loading the assemblies does not block
waiting for a draw. In that case, trying to use a create tool may fail to
create anything.

This change adds a cmds.refresh() call immediately after creating the test
window/viewport to ensure that the geometry that we're making live is present
and will work as a live surface.