cleanup attribute vs. property terminology for UsdShade inputs and outputs
UsdShade inputs and outputs are represented by UsdAttribute, not UsdProperty
(i.e. they can't be a UsdRelationship). This change cleans up function names
and doc strings to make this more clear.

This change corresponds to core USD commit
3b0c7c90f9779f955e9751c21873b637f660cd5f.

Note that the change in the underlying UsdShade API in that commit was
transparent to client code, so this change is not gated on a particular
USD_VERSION_NUM.

(Internal change: 2042555)
(Internal change: 2042852)