save and restore uniform buffer bindings during batch renderer selection
Uniform buffer bindings need to be saved and restored across Hydra Execute()
calls as a workaround to the issue described in the documentation for the
UBOBindingsSaver class. There are two such calls in the Pixar batch renderer,
one during drawing, and another in the private _TestIntersection() method which
is used when computing selections and during live surface workflows. The former
was already protected by the workaround, but the latter was not, so this change
applies the workaround there as well.