add a way for pxrUsdProxyShape to disable subpath selection with the VP2.0 render delegate
Most of the usage of pxrUsdProxyShape nodes is when they are brought in by
activating the "Collapsed" representation of pxrUsdReferenceAssembly nodes. In
that case, they are intended to be read-only proxies, and any edits to prims
within the hierarchy should be represented as assembly edits. Other workflows
that involve more direct USD authoring should favor one of the other proxy
shape node types.

This change ensures that the selection behavior of pxrUsdProxyShape nodes is
consistent between Pixar's batch renderer and the VP2.0 render delegate.