[cmake][test] fix mayaUsd_add_test on windows
for `ENV varname=varvalue` options, we were splitting varvalue on either
':' or ';', and then converting the resulting list into a platform-specific
pathsep. However, that meant that on windows we changed "C:\" to "C;\"

Instead, I changed things so that the format accepted is exactly the same
as that used for

`set_property(TEST test_name APPEND PROPERTY ENVIRONMENT ...)`

ie, values are expected to already be formatted to be platform appropriate,
and ";" values should be escaped.

I also modified behavior slightly, such that manually specified `ENV`
values are now prepended, and come before all other values. This was done
because:

- it was simpler to implement with the new changes
- it makes sense that manually specified values take highest priority,
  therefore allowing the end user to effectively pre-empt entries that
  are automatically placed later