update testPxrUsdMayaGLInstancerDraw baseline images following Nvidia driver upgrade
At Pixar, we saw subtle changes in texture sampling when the Nvidia driver
version on our systems was upgraded to 440.59, so this change updates the
baseline images for the testPxrUsdMayaGLInstancerDraw test to include those
differences.

This test is one of a handful that are currently still only run internally at
Pixar, so there should be no impact to anyone else from this change.