add Markdown version of Pixar Maya USD Plugin Doc
The user-facing documentation for Pixar's Maya USD Plugins
currently lives here:

https://graphics.pixar.com/usd/docs/Maya-USD-Plugins.html

This change adds a version of that page, cleaned up and
translated to Markdown so that it can be added to the
maya-usd repo. Since Pixar's Maya USD plugins are no
longer delivered as part of the core USD distribution,
the documentation is scheduled to be removed from the
Pixar site in the near future. Adding this Markdown
version to maya-usd preserves the information it
provided in case anyone still finds it useful.