make USD_VERSION and USD_VERSION_NUM REQUIRED_VARS in FindUSD.cmake
These CMake variables are used in many places throughout the code base, so the
build is likely to break in ways that would be difficult to decipher if they
happen to be undefined. This change marks them as required so that the failure
occurs sooner and can be debugged more easily.