remove Pixar batch renderer's support of MAYA_VP2_USE_VP1_SELECTION env var
Native Viewport 2.0 selection is supported in the USD proxy shape for Maya 2018
Update 4 and later, so there should no longer be any need to set
MAYA_VP2_USE_VP1_SELECTION=1. With the legacy viewport well down the road of
deprecation, support for it in the batch renderer is being steadily removed as
well.

(Internal change: 2058075)