[pxrUsdMayaGL] use render params useWireframe and query displayStatus when computing repr
This change renames the shape adapter GetReprSelectorForDisplayState() function
to GetReprSelectorForDisplayStyle(), makes it non-virtual, and only requires
that the displayStyle be provided. It will query Maya for the shape's
displayStatus itself.

We also use the render params' useWireframe field to determine whether or not
to use a repr that includes the wireframe, in addition to the displayStyle.

Note that we are still erroneously using the displayStyle at Sync() time to
decide whether or not to enable lighting. This can be addressed later but for
now it only causes subtly incorrect visual results for wireframe display
styles in certain cases.