[pxrUsdMayaGL] always store the wireframe color in the render params
Since we may not get another Sync() callback if the viewport display style is
changed, we need to make sure that the wireframe color is always available from
the render params.

This may lead to slightly different shape adapter bucketing behavior, since the
render params for shapes will pretty much always have a non-zero wireframe
color now, but it shouldn't practically impact performance significantly.

This also modifies the _GetWireframeColor() function to handle the MColor to
GfVec4f conversion so that clients don't have to, and so that they can pass a
pointer to the render params' wireframeColor directly.