[pxrUsdMayaGL] make shape adapter base class responsible for identifiers
This moves all computation of a unique identifier for the shape into the shape
adapter base class and removes it from derived classes. Instead of a fixed
string and a hash of the Maya object, we instead query Maya for the node's UUID
and use that directly as the shape's "identifier", which we also use both as
the name of its rprim collection, and to construct the delegate ID.

We use the node's UUID since MPxDrawOverrides may be constructed and destructed
over the course of a node's lifetime (e.g. by doing 'ogs -reset'). Since the
draw override owns the shape adapter, we therefore need an identifier tied to
the node's lifetime rather than to the shape adapter's lifetime.

With the delegate ID now stored in the base class, its accessor no longer needs
to be virtual, and also gets defined inline.