[pxrUsdMayaGL] add a way to plumb shape adapters into GetRenderTasks() using prim filters
This will allow us to delay computing which repr to use until draw time and to
query the necessary data from the shape adapters directly. The original
mechanism will be supported by specifying the shape adapter as nullptr in the
prim filter.

This commit adds the plumbing and a subsequent commit will take advantage of it
when the shape adapters manage a collection per repr.