[pxrUsdMayaGL] make shape adapters manage a collection and a render task ID per repr
This change provides a consistent one-to-one mapping between reprs, collections,
and render tasks in the shape adapter. This results in minimal draw batch
thrashing within hydra when the repr changes due to a change in Maya viewport
displayStyle, since the repr on the collection never needs to change and the
collection on the render task never needs to change.

The shape adapter must be queried for which collection and render task ID to
use based on a particular repr, so that handling has been added where necessary
in the batch renderer and scene/task delegate. When rendering, the shape
adapters are plumbed through to GetRenderTasks() where we determine what repr
to use based on the displayStyle, and then use that repr to get the appropriate
collection and render task ID for each shape we're drawing.