[pxrUsdMayaGL] push "drawShape" out of shape adapter Sync() and into GetRenderTasks()
Storing whether or not to draw the shape during Sync() is not correct since we
may be drawing for multiple viewports with differing displayStyles, or we may
change displayStyles on a particular viewport and not get another
prepareForDraw() callback.

To fix this, instead of setting the delegate's root visibility during sync, we
check during GetRenderTasks() to see whether there's an active render for the
given displayStatus, and simply skip adding the render task if there isn't one.