add a UsdPreviewSurfaceCore fragment graph that accepts transparency instead of opacity
This fragment graph is a duplicate of the original UsdPreviewSurface graph with
one modification to have it take a "transparency" value instead of an "opacity"
value.

Subsequent changes will modify the original UsdPreviewSurface fragment graph to
forward nearly all of its connections to this new core, and to make the shading
node override for pxrUsdPreviewSurface use the core directly, since it converts
"opacity" to "transparency" itself.