MAYA-105603 Selection by Kind
Defined a "UsdSelectionKind" option var to control the selection behavior.

If the option var is assigned with a non-empty string to indicate a Kind token,
the nearest ancestor that is assigned with that Kind token or a subcategory of
that Kind token will be selected when a prim is picked in the viewport. If the
prim doesn't have any qualified ancester, it falls back to the default behavior
which will select the exact prim (can be slower due to walking the hierarchy).

If the option var is undefined or empty, the default behavior will be used to
select the exact prim.

For details about Kind, please refer to "Kind : Extensible Categorization"
(https://graphics.pixar.com/usd/docs/api/kind_page_front.html#mainpage_kind).

Examples:

// Select the nearing ancester which is a model or its subcategory.
optionVar -sv "UsdSelectionKind" "model";

// Select the nearing ancester which is a group or its subcategory.
optionVar -sv "UsdSelectionKind" "group";

// Select the nearing ancester which is a user-defined kind or its subcategory.
optionVar -sv "UsdSelectionKind" "user-defined";

// Select the exact prim that is picked in the viewport.
optionVar -sv "UsdSelectionKind" "";