[mayaUsd] load Plug libraries when finding and loading in the registry
This change ensures that even if a Plug plugin does not identify a Maya plugin
to be loaded in the "mayaPlugin" field, we still instruct the Plug library to
load its plugin, since doing so might be required to cause
readers/writers/adapters/etc. to register.

This became problematic after some restructuring of the maya-usd repo such that
the translator code was moved under lib/usd/translators, which though it is
still considered "core" mayaUsd, it is built as a separate component library.
Simply importing mayaUsd in Python does *not* cause that library to be loaded,
but loading the pxrUsd Maya plugin does. The prim writers and readers are still
identified in the plugInfo.json, but there is no longer a pxrUsdTranslators
plugin on the Maya side that goes with them.

This manifested as UsdMayaAdaptors not working in Python unless you
specifically loaded the pxrUsd plugin beforehand. With this change, that
shouldn't be necessary.

To verify this, the loadPlugin("pxrUsd") command was removed from the
testUsdMayaAdaptor test to make sure that the library is loaded correctly
through the registry discovery mechanism. The one gotcha here is that the
undo/redo test *does* depend on the pxrUsd plugin being loaded because the
adaptors can be used with or without undo support, and it is only available
when the plugin *is* loaded. To deal with that, the undo/redo test was just
moved into its own test script rather than trying to carefully control when the
plugin gets loaded and having to ensure that the suite of tests run in a
specific order.

(Internal change: 2088800)