MAYA-105732 Do not auto turn on displayColors
The vertex color streams are sometimes hijacked to store custom
attribute values that are not representable as colors. This is the case
in the Kitchen_Set scene with pose information that show as black when
displayColors is turned on. Let the user manually turn that feature on
if he knows for sure there are colors to display.