migrate testPxrUsdMayaGL from pxr plugin as first example of interactive tests
This migrates the testPxrUsdMayaGL test from plugin/pxr/... into the test/...
area and uses the newly added PYTHON_SCRIPT and INTERACTIVE options in
mayaUsd_add_test() to run the test script in an interactive session of Maya.

If this approach proves viable on all platforms and when run as part of CI
systems, the rest of the pxrUsdMayaGL tests will be migrated as well.