adjust verification of texture path relative-ness in testUsdExportImportRoundtripPreviewSurface
When verifying that texture paths are authored as relative file paths when
exported to USD, this change makes that testing more robust to changes in where
the test data is located versus where the test is run from. The expectation is
that the texture path in USD is authored as a file path relative to the USD
file, but depending on where the files are located on disk, that relative path
may or may not involve a "..". Instead we check that joining the directory that
the USD file lives in with the relative texture path provides a full path to a
file that exists.