MAYA-105263 Improve draw performance for dense mesh
The face vertex indices can jump drastically and cause cache hit rate
to be low when drawing. Thus we need to sort the vertices so it can be
more cache-friendly.

After sorting, a remapping table is populated and used to prepare vertex
buffers. This remapping table is updated only when topology is changed,
so it won't happen often in common workflows.

Previously we have had another remapping code path for vertex unsharing
when there is any uniform or face-varying primvar. The commit will unify
the two remapping processes with a single remapping table, making it
easier to maintain.