use the un-linearized lambert transparency to author displayOpacity
Prior to PR #707, the un-linearized transparency value was being used to
compute a transparency "average", which was then used to author displayOpacity
on the bound gprim if it did not already have an authored displayOpacity, as
well as a displayOpacity input on the shader prim. It looks like an unintended
side effect of #707 was that it was made to use the linearized transparency for
this purpose instead.

Both of these behaviors seem somewhat dubious. It doesn't really seem like the
shading mode exporter should be authoring displayOpacity of the bound gprim,
since that is more a property of the geometry than the material. It also
doesn't appear as if the displayOpacity input being authored on the shader prim
is used for anything.

Ignoring those two oddities for now, the changes here should address the
"regression" introduced by #707. We can revisit the behavior of the
displayColors shading mode (or better yet work towards removing it!) in
follow-up commits.