make fixturesUtils.py available to all tests
This change moves the fixturesUtils.py module previously local to the
translators tests up to a "project-level" testUtils directory. The contents of
the testUtils directory is installed into the following location in the build
directory:

${CMAKE_BINARY_DIR}/test/python

We then add that directory to PYTHONPATH when registering tests so that *all*
tests will have access to the Python module, and any others we choose to add
in testUtils. This avoids polluting the install directory with test-only Python
modules.