remove Python test dependencies on future.utils.iteritems and builtins.zip
This removes usage of future.utils.iteritems from the usdMaya and translators
tests to make it easier to test for currently released versions of Maya and
Python 2. testUsdImportXforms required some additional surgery to remove the
use of builtins.zip which may also not be available in certain installs of
Python 2.

Outside of this commit, there are three remaining pxrUsdMayaGL tests that still
use future.utils.iteritems, but those are addressed in PR #761 which migrates
those tests from plugin/pxr/... to test/...

Combining the changes here with those in #761 removes all usage of the future
Python module, so the instructions on how to install it were removed from
build.md.