remove root prim xformOps for testProxyShapeLiveSurface test USD
The pxrUsdReferenceAssembly workflow never really supported significant
authoring on the root prim that is referenced by the assembly. This means that
for the purposes of the testProxyShapeLiveSurface test, the existing authoring
that attempts to translate the root prim is actually being ignored when the USD
is referenced by a USD reference assembly or proxy shape. The test was authored
taking that into account.

However, that is only the case when using Pixar's batch renderer. This issue
has been fixed when using UFE and the Viewport 2.0 render delegate. In that
case, using the same Maya scene and USD, the referenced geometry ends up being
transformed correctly, but differently as compared to the Pixar batch renderer.

To keep the test results consistent (and correct) between the Pixar batch
renderer and the VP2.0 render delegate, we simply remove the root prim
xformOps.

(Internal change: 2094559)