delete testing UI and resource handles in live surface and selection performance tests
This is an attempt to drop all handles the testProxyShapeLiveSurface and
testProxyShapeSelectionPerformance tests are holding to UI and viewport
elements in hopes that their resources can be fully freed when Maya exits.
This was raised as an issue in PR #761 with debug builds of Maya that assert
when resources are not freed at exit.