add a symmetric shader writer for literal translation of Maya shaders to USD
It turns out that there is very little specific to RenderMan for Maya being
done in the "pxrRis" shadingMode, so this change extracts the shader export
translation from there into a re-usable shader writer for use with the
"useRegistry" shadingMode.

This new shader writer performs a "literal" translation of the Maya shading
node to USD. Input and output attributes on the Maya node translate directly
to inputs and outputs on the exported UsdShadeShader with the same names. The
shader writer can optionally specify a material conversion that must be set
in order for the writer to return `Supported` from its `CanExport()` function.
If no conversion name is supplied, `CanExport()` will always return
`Supported`. If a conversion name is supplied, `Supported` is returned if the
conversion name matches the one specified in the export args, and `Unsupported`
is returned otherwise.

A static `RegisterWriter()` function is provided to simplify the registration
of writers that use this new class. `RegisterWriter()` should be called inside
a `TF_REGISTRY_FUNCTION(UsdMayaShaderWriterRegistry)` block.