add a symmetric shader reader for literal translation of USD Shaders to Maya nodes
Similar to work done on the export side, this change extracts the shader import
translation from the `pxrRis` shading mode into a re-usable shader reader for
use with the `useRegistry` shadingMode.

This new shader reader performs a "literal" translation of a USD Shader ID to
Maya shading nodes of a particular type. Values and connections on inputs
and outputs of the Shader prim are translated directly to attributes with
the same names on the Maya node.

The shader reader can optionally be restricted to a particular shading
conversion. If no shading conversion name is supplied, the reader's
`CanImport()` function will always return `Fallback`. If a shading conversion
name is supplied, `Fallback` is returned if the shading conversion name matches
the one specified in the import args, and `Unsupported` is returned otherwise.
A support level of "Fallback" is used to allow a more specialized reader for a
particular shader ID to take precedence.

A static `RegisterReader()` function is provided to simplify the registration
of readers that use this new class. `RegisterReader()` should be called inside
a `TF_REGISTRY_FUNCTION(UsdMayaShaderReaderRegistry)` block.