register symmetric shader readers for RenderMan for Maya
These new shader readers are intended to replace the import side of the
`pxrRis` shadingMode. This is work towards the ultimate goal of removing the
`pxrRis` shadingMode.

These shader readers hook into the `useRegistry` shadingMode and provide
`Fallback` support for import USD Shader prims with RenderMan for Maya shader
IDs. More specialized readers can be written and registered with stronger
support to take precedence if necessary.

The build of the RenderMan for Maya shader readers is enabled by default, but
it can be disabled by setting the build option BUILD_RFM_TRANSLATORS=OFF.