add test for importing RenderMan for Maya shaders in "useRegistry" shadingMode
This test is largely copied from the existing test that uses "pxrRis"
shadingMode.

Note that importing the test USD file results only in Maya-native nodes, so
there is no need that RenderMan for Maya be installed or loaded.

This test is added when the BUILD_RFM_TRANSLATORS build option is enabled.

Note also that the rendermanForMaya material conversion at export will create
the surface terminal output using the "ri" renderContext, but the test USD uses
the universal renderContext for now. Further work is needed in the
"useRegistry" shadingMode to be able to support finding materials in
renderContexts other than universal.