register a material conversion and symmetric shader writers for RenderMan for Maya
These new shader writers and the "rendermanForMaya" material conversion are
intended to replace the export side of the "pxrRis" shadingMode. This is work
towards the ultimate goal of removing the "pxrRis" shadingMode.

These shader writers hook into the "useRegistry" shadingMode and are active
when the "convertMaterialsTo" export argument is set to "rendermanForMaya".
Shading networks output by these writers are bound to the Material using the
"ri" renderContext.

The build of the RenderMan for Maya shader writers is enabled by default, but
it can be disabled by setting the build option BUILD_RFM_TRANSLATORS=OFF.