ensure UsdShadeMaterialBindingAPI schema is applied when authoring bindings during export
Since UsdShadeMaterialBindingAPI is an API schema, it should be applied any
time it is used for authoring. This allows querying for whether the API schema
has been authored on a prim by using UsdPrim::HasAPI(). See some additional
discussion of API schemas in the USD glossary here:

https://graphics.pixar.com/usd/docs/USD-Glossary.html#USDGlossary-APISchema

The shading export code here somewhat predates the concept of applying API
schemas, which is why handling of this is only now being added.