disable shadows on the directionalLight in the Maya scene for the testProxyShapeDrawColorAccuracy test
The Storm-based Pixar batch renderer (poorly) supports depth map shadows but not
ray-traced shadows, while the Viewport 2.0 render delegate supports both. The
directionalLight in the Maya scene for the testProxyShapeDrawColorAccuracy test
was set to use ray-traced shadows (which is the default). This meant that the
"all_lights" image written by the test would not include shadows when run with
the Pixar batch renderer, but it would include a very noisy shadow when run
with the Viewport 2.0 render delegate, since the geometry is just a very simple
cube.

Shadows are not important in the functionality being exercised by this test, so
this change simply disables shadows on the directionalLight.