add UFE contextOps for working set management (loading and unloading)
This adds contextOps items for "Load", "Load with Descendants", and "Unload"
based on the load set of the prim's stage, the loadable paths at or below that
prim's path, and whether the prim has a payload.

These items map to undoable commands that call UsdStage::Load() (with the
appropriate UsdLoadPolicy) or UsdStage::Unload() with the prim's path.

See the "Working Set Management" section of the USD docs on UsdStage for more
discussion of load and unload rules:
https://graphics.pixar.com/usd/docs/api/class_usd_stage.html