put asset contents of StandaloneScene Ball asset behind a payload
This will enable testing of the working set management UFE contextOps ("Load",
"Load with Descendants", and "Unload").

Note that the Ball asset's shading variants have been lofted to the near side
of the payload so that they can be queried for and selections can be made
without having to load the payload.