enable exportDisplayColor in testProxyShapeDrawColors test following #822
Internally, we have a post-process to the testProxyShapeDrawColors test that
diffs a render generated from the viewport against a baseline. Following the
merge of PR #822, the test began to fail since displayColor was no longer being
exported, so the color represented in the image was different. This change
turns on the exportDisplayColor option to cmds.usdExport() so that displayColor
is present in the exported USD, and the rendered image matches the baseline
again.