fix import of "useSpecularWorkflow" input on UsdPreviewSurface shaders
"useSpecularWorkflow" is an int-typed input attribute in USD, but it's a
boolean attribute in Maya, so the value needs to be type-converted during
import. The conversion was already being done for export.

This fixes an issue where the value was not being set on the Maya shading node
and an error message that looks like this was being emitted:

Runtime Error: in SetMayaAttr at line 719 of lib/mayaUsd/fileio/utils/readUtil.cpp -- Cannot set value of type 'int' on plug '<shaderName>.useSpecularWorkflow'