remove redundant HdChangeTracker::InstancerInserted() call in HdMayaDelegateCtx::InsertRprim()
HdRenderIndex::InsertInstancer() itself calls this, so there is no reason to do
so again. It has been that way since at least as far back as core USD 19.11, so
no USD API version guards are necessary.

Starting after core USD 20.11, the signature of InstancerInserted() was changed
to include the HdDirtyBits.

This change corresponds to core USD commit
https://github.com/PixarAnimationStudios/USD/commit/5a7331443a22d1a92abbbff50bab383c6db31dc3

(Internal change: 2118195)