allow specifying a specific SdfValueTypeName for the shader input to be authored
This adds an inputTypeName parameter to AuthorShaderInputFromShadingNodeAttr()
so that a specific value type name can be used when creating and authoring a
shader input. This is useful in cases where the role of the value type name may
not be discoverable strictly from inspecting the Maya attribute plug (for
example, determining that the "normalCamera" attributes of Maya shaders should
be exported as Normal3f rather than just Float3).