allow bypassing shader input authoring for Maya attributes that are unauthored
This adds an optional ignoreIfUnauthored parameter to
AuthorShaderInputFromShadingNodeAttr() so that unauthored and unconnected
attributes on Maya shader nodes can be ignored. It defaults to false so that
shader inputs are still authored regardless. Setting it to true may be
appropriate for cases where the Maya attribute and the shader input share the
same default value (for example, "incandescence" in Maya and "emissiveColor" in
UsdPreviewSurface are both black by default).