disable shading import in UsdMayaTranslatorCamera::ReadToCamera()
Code that links against mayaUsd but otherwise does not load any Maya plugins
currently encounters an error that looks like this when it calls ReadToCamera():

Error in '_shadingModesImportArgs' at line 244 in file ext/mayapkg/maya_open_source/maya-usd/lib/mayaUsd/fileio/jobs/jobArgs.cpp : 'Unknown material conversion 'UsdPreviewSurface''

The issue is that the "UsdPreviewSurface" shading conversion is specified as
the default in the default args dictionary in the mayaUsd core, but that
conversion is only registered when one of the mayaUsdPlugin, usdPreviewSurface,
or pxrUsdPreviewSurface plugins are loaded.

Since in this case we're only interested in importing cameras, we can work
around that issue for now by overriding "shadingMode" in the default args
dictionary and disabling shading import.